requirements.txt
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
qrcode==6.1
six==1.15.0

